+++ flask-app/config.py
@@ -1,0 +1,6 @@
+import os
+
+AWS_ACCESS_KEY_ID = os.getenv("AWS_ACCESS_KEY_ID")
+AWS_SECRET_ACCESS_KEY = os.getenv("AWS_SECRET_ACCESS_KEY")
+S3_URL = "https://fly.storage.tigris.dev/"
+BUCKET_NAME = "pothole-images"
