+++ flask-app/templates/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}"/>
     <!-- at top of index.html -->
-    <a href="{{ url_for('dashboard') }}">Go to Pothole Map Dashboard →</a>
+    <a href="{{ url_for('dashboard.dashboard') }}">Go to Pothole Map Dashboard →</a>
     <link rel="stylesheet" href="https://unpkg.com/leaflet/dist/leaflet.css"/>
     <title>Upload Potholes</title>
     <style>
